Adjustments components and display dates
