+++ web/package.json
@@ -20,8 +20,8 @@
     "@types/react-dom": "18.0.11",
     "autoprefixer": "10.4.14",
     "axios": "^1.3.5",
+    "dayjs": "^1.11.9",
     "install": "^0.13.0",
-    "moment": "^2.29.4",
     "moment-timezone": "^0.5.43",
     "next": "13.3.0",
     "next-redux-wrapper": "^8.1.0",
