+++ web/src/components/CardHistory.tsx
@@ -1,5 +1,5 @@
-import moment from "moment";
 import numeral from "numeral";
+import dayjs from "dayjs";
 
 import { ArrowRight } from "@phosphor-icons/react";
 
@@ -18,7 +18,7 @@ export const CardHistory = ({
   isExit,
   createdAt,
 }: CardHistoryProps) => {
-  const dateFormat = moment(createdAt).format("DD/MM/YYYY");
+  const dateFormat = dayjs(createdAt).format("DD/MM/YYYY");
 
   const valueFormat = `R$ ${numeral(value).format("0,0")}`;
