+++ web/src/lib/dayjs.ts
@@ -1,0 +1,4 @@
+import dayjs from "dayjs";
+import "dayjs/locale/pt-br";
+
+dayjs.locale("pt-br");
