+++ web/src/pages/_document.tsx
@@ -14,6 +14,7 @@ export default function Document() {
           href='https://fonts.googleapis.com/css2?family=Poppins:wght@200;300;400;500;600;700;800;900&display=swap'
           rel='stylesheet'
         />
+        <title>My Finances</title>
       </Head>
       <body className='bg-background bg-no-repeat overflow-y-hidden'>
         <Main />
